Check if recipes already installed

+++ eventstore/recipes/install.rb
@@ -1,3 +1,5 @@
+require 'win32/service'
+
 powershell_script 'install eventstore' do
   code <<-EOH
   Add-Type -assembly "system.io.compression.filesystem"
@@ -7,6 +9,7 @@ powershell_script 'install eventstore' do
   nssm install EventStore "#{node[:eventstore][:destination_path]}\\EventStore.ClusterNode.exe" --db ./db --log ./logs
   EOH
   action :run
+  not_if ::Win32::Service.exists?("EventStore")
 end
 
 service 'EventStore' do
